+++ R/app.R
@@ -526,6 +526,10 @@ new_app <- function() {
       if (!res$.sent && is.null(res$.delay)) {
         if (!res$headers_sent) {
           res$send_status(404)
+        } else if ((res$get_header("Transfer-Encoding") %||% "") == "chunked") {
+          res$
+            send_chunk(raw(0))$
+            send("")
         } else {
           res$send("")
         }