+++ src/rweb.c
@@ -29,6 +29,7 @@ SEXP response_send_headers(SEXP req);
 SEXP response_send(SEXP req);
 SEXP response_write(SEXP req, SEXP data);
 SEXP response_send_error(SEXP req, SEXP message, SEXP status);
+SEXP response_send_chunk(SEXP req, SEXP data);
 
 SEXP presser_crc32(SEXP v);
 
@@ -47,7 +48,7 @@ static const R_CallMethodDef callMethods[]  = {
   { "response_send",         (DL_FUNC) &response_send,         1 },
   { "response_write",        (DL_FUNC) &response_write,        2 },
   { "response_send_error",   (DL_FUNC) &response_send_error,   3 },
-
+  { "response_send_chunk",   (DL_FUNC) &response_send_chunk,   2 },
   /* others */
   { "presser_crc32",        (DL_FUNC) &presser_crc32,        1 },
   { NULL, NULL, 0 }
@@ -726,6 +727,28 @@ SEXP response_write(SEXP req, SEXP data) {
   fprintf(stderr, "conn %p: writing %d bytes\n", conn, len);
 #endif
   CHK(mg_write(conn, RAW(data), len));
+
+  UNPROTECT(2);
+  return R_NilValue;
+}
+
+SEXP response_send_chunk(SEXP req, SEXP data) {
+  SEXP res = PROTECT(Rf_findVar(Rf_install("res"), req));
+  SEXP headers_sent = PROTECT(Rf_findVar(Rf_install("headers_sent"), res));
+  if (! LOGICAL(headers_sent)[0]) response_send_headers(req);
+
+  SEXP xconn = Rf_findVar(Rf_install(".xconn"), req);
+  struct mg_connection *conn = R_ExternalPtrAddr(xconn);
+
+  r_call_on_early_exit(response_cleanup, conn);
+
+  int ret = 0;
+  int len = LENGTH(data);
+
+#ifndef NDEBUG
+  fprintf(stderr, "conn %p: sending chunk of %d bytes\n", conn, len);
+#endif
+  CHK(mg_send_chunk(conn, (const char*) RAW(data), len));
 
   UNPROTECT(2);
   return R_NilValue;