+++ tests/testthat/test-response.R
@@ -67,3 +67,16 @@ test_that("write-wait", {
   expect_equal(resp$status_code, 200)
   expect_equal(rawToChar(resp$content), "hello world!")
 })
+
+test_that("send_chunk", {
+  url <- web$url("/send-chunk")
+  resp <- curl::curl_fetch_memory(url)
+  expect_equal(resp$status_code, 200)
+  expect_equal(
+    rawToChar(resp$content),
+    "first chunk\nsecond chunk\nthird and final chunk\n"
+  )
+  headers <- curl::parse_headers_list(resp$headers)
+  expect_equal(headers[["content-type"]], "text/plain")
+  expect_equal(headers[["transfer-encoding"]], "chunked")
+})