+++ NEWS.md
@@ -1,6 +1,11 @@
 
 # Development version
 
+* Support chunk resposes via the new `response$send_chunk()` method.
+
+* New `httpbin_app()` end point: `/stream-bytes`, to send chunked
+  random bytes.
+
 # 1.0.0
 
 First public release.